feat: Delete old transports
Delete old transports as we've switched entirely to new v7 transports.
