+++ packages/browser/src/transports/index.ts
@@ -1,6 +1,2 @@
-export { BaseTransport } from './base';
-export { FetchTransport } from './fetch';
-export { XHRTransport } from './xhr';
-
-export { makeNewFetchTransport } from './new-fetch';
-export { makeNewXHRTransport } from './new-xhr';
+export { makeNewFetchTransport } from './fetch';
+export { makeNewXHRTransport } from './xhr';
